requirement.txt
django
django-extensions
djangorestframework
django-filter
python-dotenv
mysql-connector-python